requirements.txt
bleak==0.21.1
loguru==0.7.2
boto3==1.34.49
bluepy==1.3.0
pyserial==3.5
